+++ stream_alert_cli/utils.py
@@ -238,134 +238,6 @@ def generate_subparser(parser, name, description=None, subcommand=False):
 
 
 
-def _setup_test_subparser(subparser):
-    """Add the test subparser: manage.py test"""
-    test_subparsers = subparser.add_subparsers(dest="test subcommand", required=True)
-
-    _setup_test_classifier_subparser(test_subparsers)
-    _setup_test_rules_subparser(test_subparsers)
-    _setup_test_live_subparser(test_subparsers)
-
-
-def _setup_test_classifier_subparser(subparsers):
-    """Add the test validation subparser: manage.py test classifier [options]"""
-    test_validate_parser = generate_subparser(
-        subparsers,
-        'classifier',
-        description='Validate defined log schemas using integration test files',
-        subcommand=True
-    )
-
-    _add_default_test_args(test_validate_parser)
-
-
-def _setup_test_rules_subparser(subparsers):
-    """Add the test rules subparser: manage.py test rules [options]"""
-    test_rules_parser = generate_subparser(
-        subparsers,
-        'rules',
-        description='Test rules using integration test files',
-        subcommand=True
-    )
-
-    # Flag to run additional stats during testing
-    test_rules_parser.add_argument(
-        '-s',
-        '--stats',
-        action='store_true',
-        help='Enable outputing of statistical information on rules that run'
-    )
-
-    # Validate the provided repitition value
-    def _validate_repitition(val):
-        """Make sure the input is between 1 and 1000"""
-        err = ('Invalid repitition value [{}]. Must be an integer between 1 '
-               'and 1000').format(val)
-        try:
-            count = int(val)
-        except TypeError:
-            raise test_rules_parser.error(err)
-
-        if not 1 <= count <= 1000:
-            raise test_rules_parser.error(err)
-
-        return count
-
-    # flag to run these tests a given number of times
-    test_rules_parser.add_argument(
-        '-n',
-        '--repeat',
-        default=1,
-        type=_validate_repitition,
-        help='Number of times to repeat the tests, to be used as a form performance testing'
-    )
-
-    _add_default_test_args(test_rules_parser)
-
-
-def _setup_test_live_subparser(subparsers):
-    """Add the test live subparser: manage.py test live [options]"""
-    test_live_parser = generate_subparser(
-        subparsers,
-        'live',
-        description='Run end-to-end tests that will attempt to send alerts to each rule\'s outputs',
-        subcommand=True
-    )
-
-    _add_default_test_args(test_live_parser)
-
-
-def _add_default_test_args(test_parser):
-    """Add the default arguments to the test parsers"""
-    test_filter_group = test_parser.add_mutually_exclusive_group(required=False)
-
-    # add the optional ability to test against a rule/set of rules
-    test_filter_group.add_argument(
-        '-f',
-        '--test-files',
-        dest='files',
-        metavar='FILENAMES',
-        nargs='+',
-        help='One or more file to test, separated by spaces',
-        action=UniqueSetAction,
-        default=set()
-    )
-
-    # add the optional ability to test against a rule/set of rules
-    test_filter_group.add_argument(
-        '-r',
-        '--test-rules',
-        dest='rules',
-        nargs='+',
-        help='One or more rule to test, separated by spaces',
-        action=UniqueSetAction,
-        default=set()
-    )
-
-    # add the optional ability to change the test files directory
-    test_parser.add_argument(
-        '-d',
-        '--files-dir',
-        help='Path to directory containing test files',
-        default=DEFAULT_TEST_FILES_DIRECTORY
-    )
-
-    # Add the optional ability to log verbosely or use quite logging for tests
-    verbose_group = test_parser.add_mutually_exclusive_group(required=False)
-
-    verbose_group.add_argument(
-        '-v',
-        '--verbose',
-        action='store_true',
-        help='Output additional information during testing'
-    )
-
-    verbose_group.add_argument(
-        '-q',
-        '--quiet',
-        action='store_true',
-        help='Suppress output for passing tests, only logging if there is a failure'
-    )
 
 
 def add_default_lambda_args(lambda_parser):
@@ -430,184 +302,5 @@ def _setup_kinesis_subparser(subparser):
 
 
 
-
-
-
-
-def _setup_threat_intel_subparser(subparser):
-    """Add threat intel subparser: manage.py threat-intel [action]"""
-    _set_parser_epilog(
-        subparser,
-        epilog=(
-            '''\
-            Examples:
-
-                manage.py threat-intel \\
-                  enable \\
-                  --dynamodb-table my_ioc_table
-            '''
-        )
-    )
-
-    actions = ['disable', 'enable']
-    subparser.add_argument(
-        'action',
-        metavar='ACTION',
-        choices=actions,
-        help='One of the following actions to be performed: {}'.format(', '.join(actions))
-    )
-
-    subparser.add_argument(
-        '--dynamodb-table',
-        dest='dynamodb_table_name',
-        help='DynamoDB table name where IOC information is stored'
-    )
-
-
-def _setup_threat_intel_configure_subparser(subparsers):
-    """Add threat intel downloader configure subparser
-
-    manage.py threat-intel-downloader configure [options]
-    """
-    ti_downloader_configure_parser = _generate_subparser(
-        subparsers,
-        'configure',
-        description='Enable, disable, or configure the threat intel downloader function',
-        subcommand=True
-    )
-
-    # Enable/Disable toggle group
-    toggle_group = ti_downloader_configure_parser.add_mutually_exclusive_group(required=False)
-
-    toggle_group.add_argument(
-        '-e',
-        '--enable',
-        dest='enable_threat_intel_downloader',
-        help='Enable the threat intel downloader function',
-        action='store_true'
-    )
-
-    toggle_group.add_argument(
-        '-d',
-        '--disable',
-        dest='enable_threat_intel_downloader',
-        help='Disable the threat intel downloader function',
-        action='store_false'
-    )
-
-    # Function schedule expression (rate) arg
-    add_schedule_expression_arg(ti_downloader_configure_parser)
-
-    # Function timeout arg
-    add_timeout_arg(ti_downloader_configure_parser)
-
-    # Function memory arg
-    add_memory_arg(ti_downloader_configure_parser)
-
-    ti_downloader_configure_parser.add_argument(
-        '-r',
-        '--table-rcu',
-        help='Read capacity units to use for the DynamoDB table',
-        type=int,
-        default=10
-    )
-
-    ti_downloader_configure_parser.add_argument(
-        '-w',
-        '--table-wcu',
-        help='Write capacity units to use for the DynamoDB table',
-        type=int,
-        default=10
-    )
-
-    ti_downloader_configure_parser.add_argument(
-        '-k',
-        '--ioc-keys',
-        help='One or more IOC keys to store in DynamoDB table',
-        nargs='+',
-        action=UniqueSetAction,
-        default=['expiration_ts', 'itype', 'source', 'type', 'value']
-    )
-
-    ti_downloader_configure_parser.add_argument(
-        '-f',
-        '--ioc-filters',
-        help='One or more filters to apply when retrieving IOCs from Threat Feed',
-        nargs='+',
-        action=UniqueSetAction,
-        default=['crowdstrike', '@airbnb.com']
-    )
-
-    ti_downloader_configure_parser.add_argument(
-        '-i',
-        '--ioc-types',
-        help='One or more IOC type defined by the Threat Feed. IOC types can vary by feed',
-        nargs='+',
-        action=UniqueSetAction,
-        default=['domain', 'ip', 'md5']
-    )
-
-    ti_downloader_configure_parser.add_argument(
-        '-x',
-        '--excluded-sub-types',
-        help='IOC subtypes to be excluded',
-        action=UniqueSetAction,
-        default=['bot_ip', 'brute_ip', 'scan_ip', 'spam_ip', 'tor_ip']
-    )
-
-    ti_downloader_configure_parser.add_argument(
-        '-a',
-        '--autoscale',
-        help='Enable auto scaling for the threat intel DynamoDB table',
-        default=False,
-        action='store_true'
-    )
-
-    ti_downloader_configure_parser.add_argument(
-        '--max-read-capacity',
-        help='Maximum read capacity to use when auto scaling is enabled',
-        type=int,
-        default=5
-    )
-
-    ti_downloader_configure_parser.add_argument(
-        '--min-read-capacity',
-        help='Minimum read capacity to use when auto scaling is enabled',
-        type=int,
-        default=5
-    )
-
-    ti_downloader_configure_parser.add_argument(
-        '-u',
-        '--target-utilization',
-        help=(
-            'Target percentage of consumed provisioned throughput at a point in time '
-            'to use for auto-scaling the read capacity units'
-        ),
-        type=int,
-        default=70
-    )
-
-
-def _setup_threat_intel_auth_subparser(subparsers):
-    """Add threat intel downloader update-auth subparser
-
-    manage.py threat-intel-downloader update-auth
-    """
-    _generate_subparser(
-        subparsers,
-        'update-auth',
-        description='Enable, disable, or configure the threat intel downloader function',
-        subcommand=True
-    )
-
-
-def _setup_threat_intel_downloader_subparser(subparser):
-    """Add threat intel downloader subparser: manage.py threat-intel-downloader [subcommand]"""
-    ti_subparsers = subparser.add_subparsers(dest="threat-intel-downloader subcommand",
-                                             required=True)
-
-    _setup_threat_intel_configure_subparser(ti_subparsers)
-    _setup_threat_intel_auth_subparser(ti_subparsers)
